Fork

+++ database/migrations/2020_01_28_100002_create_received_repayments_table.php
@@ -15,7 +15,7 @@ class CreateReceivedRepaymentsTable extends Migration
     {
         Schema::create('received_repayments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('loan_id');
+            $table->unsignedBigInteger('loan_id');
 
             // TODO: Add missing columns here
